+++ plugins/gui/src/user_action/action_set_object_data.cpp
@@ -14,13 +14,19 @@ namespace hal
         return new ActionSetObjectData;
     }
 
-    ActionSetObjectData::ActionSetObjectData(QString category, QString key, QString type, QString val)
-    : mKey(key), mCat(category), mType(type), mVal(val), mOldKey(""), mOldCat(""), mKeyOrCatModified(false)
-    {
-    }
+    ActionSetObjectData::ActionSetObjectData(ActionType act, const DataTableEntry& dataEntry)
+        : mAct(act), mDataEntry(dataEntry)
+    {;}
+
+    /*
+        return DataTableEntry::fromContainer(container, mDataEntry.category.toStdString(), mDataEntry.key.toStdString());
+        return DataTableEntry::fromContainer(container, mOldCat.toStdString(), mOldKey.toStdString());
+    */
 
     bool ActionSetObjectData::exec()
     {
+        if (mDataEntry.isNull()) return false;
+
         //for now, this action only supports setting the data of gates, modules and nets (DataContainer)
         DataContainer* container = nullptr;
         switch(mObject.type())
@@ -34,38 +40,64 @@ namespace hal
             case UserActionObjectType::Net:
                 container = dynamic_cast<DataContainer*>(gNetlist->get_net_by_id(mObject.id()));
                 break;
-            default: break;
-        }
-
-        if(!container)
-            return false;
-
-        //implicit modifying of an existing entry, but it must be deleted beforehand and undo action must be constructed in an appropriate way
-        if(mKeyOrCatModified)
-        {
-            if(!container->has_data(mOldCat.toStdString(), mOldKey.toStdString()))//additional security check
+            default:
                 return false;
-
-            //construct appropriate undo action
-            auto data = container->get_data(mOldCat.toStdString(), mOldKey.toStdString());
-            mUndoAction = new ActionSetObjectData(mOldCat, mOldKey, QString::fromStdString(std::get<0>(data)), QString::fromStdString(std::get<1>(data)));
-            mUndoAction->setObject(mObject);
-            static_cast<ActionSetObjectData*>(mUndoAction)->setChangeKeyAndOrCategory(mCat, mKey);
-
-            //delete old, add new
-            container->delete_data(mOldCat.toStdString(), mOldKey.toStdString());
-            container->set_data(mCat.toStdString(), mKey.toStdString(), mType.toStdString(), mVal.toStdString());
         }
-        //check if the data to set already exists (undo action is then "change back")
-        else if(container->has_data(mCat.toStdString(), mKey.toStdString()))
+
+        ActionSetObjectData* undoAction = nullptr; // undo action will allways be of same type as action
+
+        std::string searchCategory = (mAct == MoveAction) ? mOldCat.toStdString() : mDataEntry.category.toStdString();
+        std::string searchKey      = (mAct == MoveAction) ? mOldKey.toStdString() : mDataEntry.key.toStdString();
+        DataTableEntry currentEntry = DataTableEntry::fromContainer(container, searchCategory, searchKey);
+
+        //construct appropriate undo action
+        switch (mAct)
         {
-            auto data = container->get_data(mCat.toStdString(), mKey.toStdString());
-            mUndoAction = new ActionSetObjectData(mCat, mKey, QString::fromStdString(std::get<0>(data)), QString::fromStdString(std::get<1>(data)));
-            mUndoAction->setObject(mObject);
-            container->set_data(mCat.toStdString(), mKey.toStdString(), mType.toStdString(), mVal.toStdString());
+            case UnknownAction:
+                return false;
+            case DeleteAction:
+                if (currentEntry.isNull()) return false;
+                undoAction = new ActionSetObjectData(CreateAction, currentEntry);
+                break;
+            case CreateAction:
+                undoAction = new ActionSetObjectData(DeleteAction, mDataEntry);
+                break;
+            case ModifyAction:
+                if (currentEntry.isNull())
+                    // does not exist yet, it is a create
+                    undoAction = new ActionSetObjectData(DeleteAction, mDataEntry);
+                else
+                    undoAction = new ActionSetObjectData(ModifyAction, currentEntry);
+                break;
+            case MoveAction:
+                if (currentEntry.isNull()) return false;
+                undoAction = new ActionSetObjectData(MoveAction, currentEntry);
+                undoAction->setChangeKeyAndOrCategory(mDataEntry.category, mDataEntry.key);
+                break;
         }
-        else //if the data does not already exists -> it is new data -> undo action deletes it (yet to come)
+
+        if (undoAction)
         {
+            undoAction->setObject(mObject);
+            mUndoAction = undoAction;
+        }
+
+        // modify container
+        switch (mAct)
+        {
+            case DeleteAction:
+                container->delete_data(mDataEntry.category.toStdString(), mDataEntry.key.toStdString());
+                break;
+            case CreateAction:
+            case ModifyAction:
+                mDataEntry.toContainer(container);
+                break;
+            case MoveAction:
+                container->delete_data(mOldCat.toStdString(), mOldKey.toStdString());
+                mDataEntry.toContainer(container);
+                break;
+            default:
+                break;
         }
 
         if (mObject.type() == UserActionObjectType::Gate)
@@ -82,11 +114,13 @@ namespace hal
 
     void ActionSetObjectData::writeToXml(QXmlStreamWriter &xmlOut) const
     {
-        xmlOut.writeTextElement("category", mCat);
-        xmlOut.writeTextElement("key", mKey);
-        xmlOut.writeTextElement("type", mType);
-        xmlOut.writeTextElement("value", mVal);
-        if(mKeyOrCatModified)
+        const QStringList actionTypeToString = {"unknown", "delete", "create", "modify", "move"};
+        xmlOut.writeTextElement("action",   actionTypeToString[mAct]);
+        xmlOut.writeTextElement("category", mDataEntry.category);
+        xmlOut.writeTextElement("key",      mDataEntry.key);
+        xmlOut.writeTextElement("type",     mDataEntry.dataType);
+        xmlOut.writeTextElement("value",    mDataEntry.value);
+        if(mAct == MoveAction)
         {
             xmlOut.writeTextElement("oldCategory", mOldCat);
             xmlOut.writeTextElement("oldKey", mOldKey);
@@ -95,40 +129,43 @@ namespace hal
 
     void ActionSetObjectData::readFromXml(QXmlStreamReader &xmlIn)
     {
+        const QStringList actionTypeToString = {"unknown", "delete", "create", "modify", "move"};
         while(xmlIn.readNextStartElement())
         {
             if(xmlIn.name() == "category")
-                mCat = xmlIn.readElementText();
+                mDataEntry.category = xmlIn.readElementText();
             else if(xmlIn.name() == "key")
-                mKey = xmlIn.readElementText();
+                mDataEntry.key      = xmlIn.readElementText();
             else if(xmlIn.name() == "type")
-                mType = xmlIn.readElementText();
+                mDataEntry.dataType = xmlIn.readElementText();
             else if(xmlIn.name() == "value")
-                mVal = xmlIn.readElementText();
+                mDataEntry.value    = xmlIn.readElementText();
             else if(xmlIn.name() == "oldCategory")
-            {
-                mKeyOrCatModified = true;
                 mOldCat = xmlIn.readElementText();
-            }
             else if(xmlIn.name() == "oldKey")
-            {
-                mKeyOrCatModified = true;
                 mOldKey = xmlIn.readElementText();
+            else if (xmlIn.name() == "action")
+            {
+                QString tmpAct = xmlIn.readElementText();
+                int i = actionTypeToString.indexOf(tmpAct);
+                mAct = i < 0 ? UnknownAction : (ActionType) i;
             }
         }
     }
 
     void ActionSetObjectData::addToHash(QCryptographicHash &cryptoHash) const
     {
+        char ctype = (char) mAct;
+        cryptoHash.addData(&ctype, 1);
         cryptoHash.addData("cat", 3);
-        cryptoHash.addData(mCat.toUtf8());
+        cryptoHash.addData(mDataEntry.category.toUtf8());
         cryptoHash.addData("key", 3);
-        cryptoHash.addData(mKey.toUtf8());
+        cryptoHash.addData(mDataEntry.key.toUtf8());
         cryptoHash.addData("type", 4);
-        cryptoHash.addData(mType.toUtf8());
+        cryptoHash.addData(mDataEntry.dataType.toUtf8());
         cryptoHash.addData("val", 3);
-        cryptoHash.addData(mVal.toUtf8());
-        if(mKeyOrCatModified)
+        cryptoHash.addData(mDataEntry.value.toUtf8());
+        if(mAct == MoveAction)
         {
             cryptoHash.addData("oldCat", 6);
             cryptoHash.addData(mOldCat.toUtf8());
@@ -139,9 +176,9 @@ namespace hal
 
     void ActionSetObjectData::setChangeKeyAndOrCategory(QString oldCategory, QString oldKey)
     {
+        mAct = MoveAction;
         mOldKey = oldKey;
         mOldCat = oldCategory;
-        mKeyOrCatModified = true;
     }
 
 }